gui resize enhanced
